+++ Toggly.FeatureManagement.Nuxt/nuxt-toggly-core/build.config.ts
@@ -1,10 +1,17 @@
 import { defineBuildConfig } from 'unbuild'
 
 export default defineBuildConfig({
-  entries: ['./src/index'],
+  entries: [
+    './src/index',
+    {
+      input: './src/telemetry/grpc-clients',
+      name: 'telemetry/grpc',
+    },
+  ],
   declaration: true,
   clean: true,
   rollup: {
     emitCJS: true,
   },
+  externals: ['@grpc/grpc-js', '@grpc/proto-loader'],
 })